+++ src/main/kotlin/reservation/project/domain/academy/entity/Category.kt
@@ -8,7 +8,7 @@ data class Category(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    val categoryId: Int = 0,
+    val categoryId: Long = 0,
 
     @Column(name = "category_name", nullable = false)
     val categoryName: String
